split off state transtions test
